#30 fix : Base 에서 NetworkState 관리하는 코드 수정
