+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/base/BaseActivityViewModel.kt
@@ -18,9 +18,10 @@ abstract class BaseActivityViewModel (
 
     init{
         networkManager.startNetwork()
+        checkNetworkState()
     }
 
-    fun checkNetworkState(){
+    private fun checkNetworkState(){
         viewModelScope.launch {
 
             networkManager.isNetworkConnected.collectLatest{
